Add alembic files
After running the command `alembic init phiphi/migration` and fixing
linting errors.

It was decided that it would be easier to have the migrations in the
package rather then in the root of the package. IE in
`phiphi/phiphi/migrations` rather then `phiphi/alembic`. This is because
of:
1. If it is not in the package ie in `projects/phiphi/alembic` we need
   to edit the `pyproject.toml` with the below edits or `pip install -e
   .` does not work.
2. In a later commit we are going to use `phiphi.core.config.settings`
   to set the db url. As such it would then make sense that the alembic
   is part of the package to.
3. This is how the fastAPI project generator also has it:
https://github.com/tiangolo/full-stack-fastapi-template/tree/master/backend/app

Changes that would be needed to `pyproject.toml`
[tool.setuptools.packages.find]
exclude = ["alembic*"]