Add `make fix_local_permissions` to phiphi/Makefile
This is needed to be able to edit the files generated by alembic
It is also possible to fix this with making changing the user in the
docker container. This is done by editing the Dockerfile and can add
a lot of complexity. Using `make fix_local_permissions` is a quick fix.