Fix healthcheck on phiphi postgres
There is a stupid problem that if the healthcheck is not with postgres
we get an error in the docker compose logs the hole time.

This stackoverflow seems to suggest that you can use a different user
but this didn't work for me:
https://stackoverflow.com/a/52776597/4874363

As such changing it to `postgres` seemed a simple solution that is fine
for development.