Remove schemas from phiphi
Includes removing the migrations with the schemas

This has been done because we starting having problems with alembic and
it "seeing" the tables in the schema `platform`. My feeling is that this
change is adding quiet a bit of difficult functionality that we are
currently not using. If we use schemas in the future we can look at this
functionality then.