Add GatherResponse.latest_job_run
Since an end user will almost always want to get the latest job run for
a gather I felt that it would be better to get it during the requests
for a gather rather then an end user or console having to look for the
job run.

Benjamin's thinking was that:
- The console has always get the latest job run for a gather for the
  list view. As based on the status is it going to show the correct
  actions.
- To reduce the amount of network usage needed we then get the latest
  job run on every gather GET

Reducing the amount of network usage will help end users with slow
internet.

My expectation is that this will not slow the gathers get down in a
heavily noticeable way. If we notice that it becomes very
slow because it has to make a query to the db for each gather we can
revisit if it is better to remove `latest_job_run` from gathers.