Add ChildType Enum to GatherResponse
The idea is that this child type can then be used thoughout the code
base.

An Enum was chosen over a Literal as a in my mind: Literal Types cannot
include values that are not literal. This means variables or expressions
evaluated at runtime are not permissible as Literal Types. ChildType
will be computed from the data in the database and will be passed to
functions at run time.