Phiphi: refactor PhiphiJobType to job_runs.schemas.ForeignJobType
Removing the literal means that python code that is going to call the
function has to pass an enum. It does however remove the duplication of
the list of types for flow_runner_flow and seems to make sense in the
code.

The disadvantage of this is that the flow_runner_flow is more coupled
with the job_runs.