Phiphi: refactor delete gather model to delete_job_run
It is thought that using a direct one to one relationship in the model
would be more transparent about how the delete of a gather works. It
also allows for the console to give the status of a delete_job_run
easily.

Also here the filter for deleted records has been removed. This will be
revisited in a later commit.