Phiphi: refactor pipeline gather.[platform|data_type]
For project db tables gather_batches and generalised_messages:
- Removes the columns `source`
- Has `platform`, `data_type` as constants in the code rather then from
  the schema
- adds `gather_type`

These changes mean that we can take source, platform and data_type out
of the gather model. This is an improvement as is also for more
transparent code for these columns and the gathers.

Rename of GatherCreationDefaults in the next commit.