Phiphi: Update pandera generalised messages schema
While there are sources where these are Int's. The generalised messages table should be the common denominator among all the data; we don't want ints to come in from tiktok. After the datatype in Bigquery becomes int, and Strings are inserted from a different source, we'll get errors.

This is also needed for using pandera.Schema.example(n) to generate sample data using the schema.